refactor(pptx): extract fill.rs from lib.rs (verbatim)
Fill/colour/stroke/effect/3D-scene/custom-geometry parsing: parse_color_node, parse_color_node_tint, parse_fill, parse_blip_fill, parse_blip_alpha, parse_fill_rect, parse_tile, parse_arrow_end, parse_stroke, parse_shadow, parse_inner_shadow, parse_shadow_node, parse_glow, parse_soft_edge, parse_reflection, EffectLst, parse_effect_lst, parse_rot3d, parse_scene3d, parse_bevel3d, parse_sp3d, parse_path_cmd, parse_cust_geom, parse_xfrm, parse_background, parse_table_style_fill — now pub(crate). fill.rs imports PptxSchemeResolver from theme, mime_from_ext from ooxml_common, and the shared XML helpers (child/children_vec/attr/attr_r/attr_i64/attr_f64) which are now pub(crate) in lib.rs (R_NS too). lib.rs uses fill::*.
